pyproject.toml
[project]
name = "ai-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "groq>=1.0.0",
    "logfire>=4.21.0",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pypdf>=6.6.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.0",
    "tavily-python>=0.7.21",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
]
